game doesn't crash anymore when a cockpit gets destroyed, atm the player disconnects and has to start again

+++ BasicGameNetwork/src/netclient/ClientNetMsgListener.java
@@ -93,7 +93,7 @@ public class ClientNetMsgListener implements MessageListener<Client> {
                     }
                     
                     // add testbox
-                    Spatial spatial;
+                    /*Spatial spatial;
                     Material material;    
                     
                     Box box = new Box(1, 0.4f, 1);
@@ -106,7 +106,7 @@ public class ClientNetMsgListener implements MessageListener<Client> {
                     material.setColor("Diffuse", color);
                     
                     spatial.setMaterial(material);
-                    ship.shipRoot.attachChild(spatial);
+                    ship.shipRoot.attachChild(spatial);*/
                     ship.shipRoot.setLocalTranslation(msg.getPos());
                     
                     return null;
@@ -120,7 +120,7 @@ public class ClientNetMsgListener implements MessageListener<Client> {
             
             this.app.enqueue(new Callable() {
                 public Object call() throws Exception {
-                    if(app.gameRunState.playerShip.id == msg.getId()) {
+                    if(app.gameRunState.playerShip != null && app.gameRunState.playerShip.id == msg.getId()) {
                         app.gameRunState.nearStation = msg.getNearby();
                     }
                     return null;
@@ -206,8 +206,11 @@ public class ClientNetMsgListener implements MessageListener<Client> {
             
             this.app.enqueue(new Callable() {
                 public Object call() throws Exception {
-                    app.gameRunState.graphicObjects.get(msg.getId()).update(msg);
-                    return null;
+                    if (app.gameRunState.graphicObjects.get(msg.getId()) != null) {
+                        app.gameRunState.graphicObjects.get(msg.getId()).update(msg);
+                    }
+                    
+                    return null;                    
                 }
             });
         } else if (m instanceof DeleteGraphicObjectMsg) {
@@ -215,7 +218,9 @@ public class ClientNetMsgListener implements MessageListener<Client> {
             
             this.app.enqueue(new Callable() {
                 public Object call() throws Exception {
-                    app.gameRunState.graphicObjects.get(msg.getId()).delete();
+                    if (app.gameRunState.graphicObjects.get(msg.getId()) != null) {
+                        app.gameRunState.graphicObjects.get(msg.getId()).delete();
+                    }
                     return null;
                 }
             });
